WIP: coverage working much better, try tests/functional/validation.py::test_python_versions
