+++ .activate.sh
@@ -1,2 +1,3 @@
+export TOP=$PWD
 make venv
 . venv-venv_update/bin/activate
