+++ tests/testing/__init__.py
@@ -104,3 +104,15 @@ def pip_freeze(venv='virtualenv_run'):
 
     assert err == ''
     return out
+
+
+def enable_coverage(tmpdir):
+    venv = tmpdir.join('virtualenv_run')
+    if not venv.isdir():
+        run('virtualenv', venv.strpath)
+    run(
+        venv.join('bin/python').strpath,
+        '-m', 'pip.__main__',
+        'install',
+        '-r', TOP.join('requirements.d/coverage.txt').strpath,
+    )
